+++ fence/models.py
@@ -42,7 +42,8 @@ from userdatamodel.models import (
     User)
 
 from fence import logger
-from fence.config import config
+#todo: models relies on config, config relies on models!!!
+# from fence.config import config
 from fence.errors import UserError
